+++ v2/bin/fig2.jl
@@ -289,10 +289,10 @@ plot!(p[13],xticks=(3.0:0.5:4.0,["3","3.5","4"]),bottom_margin=5.0mm)
 plot!(p[14],xticks=(2:1:5,["","3", "4","5"]),bottom_margin=5.0mm)
 plot!(p[15],xticks=(4:2:9,["4","6","8"]),bottom_margin=5.0mm)
 plot!(p[16],xticks=(3:1:6,["3","4","5","6"]),bottom_margin=5.0mm)
-plot!(p[1],yticks=(2:2:10,["","4","6","8","10"]),left_margin=5.0mm)
-plot!(p[5],yticks=(2:2:6,["2","4","6"]),left_margin=5.0mm)
-plot!(p[9],yticks=(4:2:9,["4","6","8"]),left_margin=5.0mm)
-plot!(p[13],yticks=(2:2:7,["2","4","6"]),left_margin=5.0mm)
+plot!(p[1],yticks=(2:2:10,["","4","6","8","10"]),left_margin=7.0mm)
+plot!(p[5],yticks=(2:2:6,["2","4","6"]),left_margin=7.0mm)
+plot!(p[9],yticks=(4:2:9,["4","6","8"]),left_margin=7.0mm)
+plot!(p[13],yticks=(2:2:7,["2","4","6"]),left_margin=7.0mm)
 
 annotate!(p[1],2.75+1.75*0.8,0.8*3.0,text("Counts",14))
 plot!(p[6],ylims=(0.0,1))
