+++ v2/bin/fig9.jl
