+++ v2/do.sh
@@ -36,32 +36,53 @@ JULIA='julia '
  #$JULIA bin/PDFfit.jl -s 45 -p  Dirichlet -d simulation-Dirichlet-44 -n 100000  --priorshift=2 &
  #$JULIA bin/PDFfit.jl -s 45 -p  Valence -d simulation-Valence-42 -n 100000 &
 
-#wait $(jobs -p)
+wait $(jobs -p)
 
-# $JULIA bin/chi2plots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
- $JULIA bin/xfxplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
-# $JULIA bin/bernsteinplots.jl -s 47 -p  Bernstein -d simulation-Bernstein-42 -f fit-Bernstein-0-45-simulation-Bernstein-42 &
-#   $JULIA bin/valenceplots.jl -s 47 -p  Valence -d simulation-Valence-42 -f fit-Valence-0-45-simulation-Valence-42  &
-#     $JULIA bin/cornerplot.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
- 
-       #$JULIA bin/gofplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
-       #$JULIA bin/gofplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
+        $JULIA bin/fig8.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
+        $JULIA bin/fig567.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
+        $JULIA bin/fig2.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  & 
+        $JULIA bin/fig9.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
 
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-1-45-simulation-Dirichlet-42 --priorshift=1  &
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-2-45-simulation-Dirichlet-42 --priorshift=2 &
+       $JULIA bin/fig34ddirichlet.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-43  &
+       $JULIA bin/fig34dirichlet.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-43  &
 
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-43  &
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-1-45-simulation-Dirichlet-43 --priorshift=1  &
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-2-45-simulation-Dirichlet-43 --priorshift=2 &
+       $JULIA bin/fig34dirichlet.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
+       $JULIA bin/fig34ddirichlet.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-0-45-simulation-Dirichlet-42  &
+       
+       $JULIA bin/fig34dirichlet.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-1-45-simulation-Dirichlet-42 --priorshift=1  &
+       $JULIA bin/fig34dirichlet.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-42 -f fit-Dirichlet-2-45-simulation-Dirichlet-42 --priorshift=2 &
 
+       $JULIA bin/fig34bernstein.jl -s 47 -p  Bernstein -d simulation-Bernstein-42 -f fit-Bernstein-0-45-simulation-Bernstein-42 &
+       $JULIA bin/fig34valence.jl   -s 47 -p  Valence -d simulation-Valence-42 -f fit-Valence-0-45-simulation-Valence-42  &
 
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-44 -f fit-Dirichlet-0-45-simulation-Dirichlet-44  &
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-44 -f fit-Dirichlet-1-45-simulation-Dirichlet-44 --priorshift=1  &
-#       $JULIA bin/allplots.jl -s 47 -p  Dirichlet -d simulation-Dirichlet-44 -f fit-Dirichlet-2-45-simulation-Dirichlet-44 --priorshift=2 &
  
  wait $(jobs -p)
+
+
+mkdir -p figures2
+
+
+cp figures/fig2-corner-fit-Dirichlet-0-45-simulation-Dirichlet-42_v3.pdf  figures2/fig2.pdf
+
+cp figures/fig3-d-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig3a.pdf
+cp figures/fig3-d-fit-Dirichlet-0-45-simulation-Dirichlet-43_v2.pdf figures2/fig3b.pdf
+cp figures/fig3-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf  figures2/fig3c.pdf
+cp figures/fig3-fit-Dirichlet-0-45-simulation-Dirichlet-43_v2.pdf  figures2/fig3d.pdf
+
+
+cp figures/fig3-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf   figures2/fig4a.pdf
+cp figures/fig3-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf   figures2/fig4b.pdf
+cp figures/fig4-fit-Bernstein-0-45-simulation-Bernstein-42_v2.pdf  figures2/fig4c.pdf
+cp figures/fig4-fit-Valence-0-45-simulation-Valence-42_v2.pdf      figures2/fig4d.pdf
+
+cp figures/fig5-momentum-corr-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig5.pdf
+cp figures/'fig6-parton-xf(x)-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf' figures2/fig6.pdf
+cp figures/fig7-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig7.pdf
+cp figures/fig8-chisq-pvalue-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig8.pdf
+cp figures/fig9-Data-GoF-fit-Dirichlet-0-45-simulation-Dirichlet-42.pdf figures2/fig9.pdf
+
+
+  echo "OK, DONE!!!"
  
- echo "OK, DONE!!!"
  
  #~/.julia/packages/PartonDensity/VMoA4/src/parametrisations/bernstein.jl
