--- v2/rename.sh
@@ -1,40 +1,0 @@
-#!/bin/bash
-set -x
-mkdir -p figures2
-
-
-cp figures/fig2-corner-fit-Dirichlet-0-45-simulation-Dirichlet-42_v3.pdf  figures2/fig2.pdf
-
-cp figures/fig3-d-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig3a.pdf
-cp figures/fig3-d-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig3b.pdf
-#cp figures/fig3-d-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf 
-#cp figures/fig3-d-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig3-d-fit-Dirichlet-2-45-simulation-Dirichlet-44_v2.pdf
-
-cp figures/fig3-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf  figures2/fig3c.pdf
-cp figures/fig3-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf  figures2/fig3d.pdf
-
-
-cp figures/fig3-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf   figures2/fig4a.pdf
-cp figures/fig3-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf   figures2/fig4b.pdf
-#cp figures/fig3-fit-Dirichlet-2-45-simulation-Dirichlet-44_v2.pdf
-cp figures/fig4-fit-Bernstein-0-45-simulation-Bernstein-42_v2.pdf  figures2/fig4c.pdf
-cp figures/fig4-fit-Valence-0-45-simulation-Valence-42_v2.pdf      figures2/fig4d.pdf
-
-cp figures/fig5-momentum-corr-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig5.pdf
-#cp figures/fig5-momentum-corr-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig5-momentum-corr-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig5-momentum-corr-fit-Dirichlet-2-45-simulation-Dirichlet-44_v2.pdf
-cp figures/'fig6-parton-xf(x)-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf' figures2/fig6.pdf
-#cp figures/'fig6-parton-xf(x)-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf'
-#cp figures/'fig6-parton-xf(x)-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf'
-#cp figures/'fig6-parton-xf(x)-fit-Dirichlet-2-45-simulation-Dirichlet-44_v2.pdf'
-cp figures/fig7-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig7.pdf
-#cp figures/fig7-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig7-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig7-fit-Dirichlet-2-45-simulation-Dirichlet-44_v2.pdf
-cp figures/fig8-chisq-pvalue-fit-Dirichlet-0-45-simulation-Dirichlet-42_v2.pdf figures2/fig8.pdf
-#cp figures/fig8-chisq-pvalue-fit-Dirichlet-1-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig8-chisq-pvalue-fit-Dirichlet-2-45-simulation-Dirichlet-42_v2.pdf
-#cp figures/fig8-chisq-pvalue-fit-Dirichlet-2-45-simulation-Dirichlet-44_v2.pdf
-cp figures/fig9-Data-GoF-fit-Dirichlet-0-45-simulation-Dirichlet-42.pdf figures2/fig9.pdf
